+++ ssdp-rs/src/lib.rs
@@ -1,5 +1,5 @@
 #![cfg_attr(all(test, unstable_assert_matches), feature(assert_matches))]
 #![cfg_attr(unstable_let_chains, feature(let_chains))]
 
-pub mod listener;
 pub mod message;
+pub mod udp;
